[Transform] Fixing types in translation

+++ src/transformations/transformer.py
@@ -3,7 +3,7 @@ from matrix.matrix import Matrix
 class Transformer():
 
     @staticmethod
-    def translation(x: int, y: int, z: int) -> Matrix:
+    def translation(x: float, y: float, z: float) -> Matrix:
         translation_matrix = Matrix.generate_identity_matrix(4)
         translation_matrix.set(0, 3, x)
         translation_matrix.set(1, 3, y)
